seperate into components
